% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data_final.R
\name{access_state_year}
\alias{access_state_year}
\title{Retrieves election data for Wisconsin from the years 2000-2020}
\usage{
access_state_year(year, data)
}
\arguments{
\item{year}{A character vector identifying the requested year}

\item{data}{the primary data frame with all of the information for the given year}
}
\value{
A list of eleven data frames, each of election data from Wisconsin and
year with the following columns
\itemize{
\item county
\item ward
\item office
\item district
\item total.votes
\item party
\item candidate
\item votes
\item contest_dem
\item contest_rep
\item year
}

Note that many of the fields may be an empty string
}
\description{
If given Wisconsin and a year, this function returns all
election data from the OpenElections github for that state and year
}
