% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/districts_baseline_func.R
\name{baseline_function}
\alias{baseline_function}
\title{baseline function}
\usage{
baseline_function(x)
}
\arguments{
\item{x}{the primary data frame with all of the information as returned by generate data}
}
\description{
baseline function
}
