% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleaning_uncontested.R
\name{candidate_function}
\alias{candidate_function}
\title{candidate_function should only be run on two party races, and summarizes the data to provide vote shares and
proportions for candidates. This allows us to then find average proportions to estimate vote shares.}
\usage{
candidate_function(x, ...)
}
\description{
candidate_function should only be run on two party races, and summarizes the data to provide vote shares and
proportions for candidates. This allows us to then find average proportions to estimate vote shares.
}
