% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleaning_uncontested.R
\name{district_func}
\alias{district_func}
\title{district_func binds together the district-based race with the statewide race to extract how that district
behaved in a statewide race. This is used to create the baseline and estimate district behavior.}
\usage{
district_func(x, y, ...)
}
\description{
district_func binds together the district-based race with the statewide race to extract how that district
behaved in a statewide race. This is used to create the baseline and estimate district behavior.
}
