% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efficiency_gap.R
\name{efficiency_gap_contested}
\alias{efficiency_gap_contested}
\title{The only difference between the efficiency gap and the efficiency gap contested
is the additional line of code before we use the object of year_baseline_data, where
we filter it to exclude uncontested estimates, which impacts the information pulled
by the function.}
\usage{
efficiency_gap_contested(full_votes, year)
}
\arguments{
\item{full_votes}{A data frame of length 99, for each legislative district and
and how they voted Democrat or Republican for a given year.}

\item{year}{A character vector identifying the requested year. The user can
input a numeric or integer vector, but the function will transform it into a character.}
}
\value{
A numeric vector, between -0.25 and 0.25 that estimates how gerrymandered the
state assembly is.
}
\description{
The only difference between the efficiency gap and the efficiency gap contested
is the additional line of code before we use the object of year_baseline_data, where
we filter it to exclude uncontested estimates, which impacts the information pulled
by the function.
}
