% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleaning_uncontested.R
\name{filter_statewide}
\alias{filter_statewide}
\title{filter_statewide takes the entire state's voting data for a given year and filters it so that only the most
prominent statewide offices are included: Senate, President, Attorney General, Secretary of State, and
Governor. In any given year, there will be some of, but never all, of these races, as they are temporally
staggered.}
\usage{
filter_statewide(x)
}
\arguments{
\item{x}{data frame created by open_elections_factory and generate_data}
}
\description{
filter_statewide takes the entire state's voting data for a given year and filters it so that only the most
prominent statewide offices are included: Senate, President, Attorney General, Secretary of State, and
Governor. In any given year, there will be some of, but never all, of these races, as they are temporally
staggered.
}
