% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data_final.R
\name{generate_data}
\alias{generate_data}
\title{Retrieves election data for Wisconsin from the years 2000-2020}
\usage{
generate_data(oe_data)
}
\arguments{
\item{state}{A character vector representing either WI or MI}

\item{year}{A character vector identifying the requested year}
}
\value{
A list of eleven data frames, each of election data from Wisconsin and
year with the following columns
\itemize{
\item county
\item ward
\item office
\item district
\item total.votes
\item party
\item candidate
\item votes
\item contest_dem
\item contest_rep
\item year
}

Note that many of the fields may be an empty string
}
\description{
If given Wisconsin and a year, this function returns all
election data from the OpenElections github for that state and year
}
