% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/un_contested_di_final.R
\name{sa_contest_all}
\alias{sa_contest_all}
\title{This makes use of the data generated in generate_data_final and the
access_state_year functiond defined there to generate a list of 11 elements
with every even year's election data on State assembly}
\usage{
sa_contest_all(data)
}
\arguments{
\item{data}{the full list of data frames from generate data}
}
\value{
an eleven element list of the major election years from 2000-2020.
Each year returns a list of two dataframes with the contested and
uncontested districts for state assembly
Each dataframe has the following columns
\itemize{
\item county
\item ward
\item office (always state assembly)
\item district
\item total.votes
\item party
\item candidate
\item votes
\item contest_dem
\item contest_rep
\item year
\item contested
}
}
\description{
Each year consists of a list of the contested districts and uncontested
districts for the state assembly election
}
\details{
currently the function only works for Wisconsin
}
