% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleaning_uncontested.R
\name{statewide_master}
\alias{statewide_master}
\title{The functions in this file are called within districts_baselin_func.R to create the baselines for uncontested
districts.
statewide_master is a function that executes smaller functions to find the total votes cast per office,
and the total two-party votes cast per office, for all statewide races in a given year. A two-party vote is
defined as the total votes cast for either the Democrat or Republican parties.}
\usage{
statewide_master(x)
}
\arguments{
\item{x}{data frame created by open_elections_factory and generate_data}
}
\description{
The functions in this file are called within districts_baselin_func.R to create the baselines for uncontested
districts.
statewide_master is a function that executes smaller functions to find the total votes cast per office,
and the total two-party votes cast per office, for all statewide races in a given year. A two-party vote is
defined as the total votes cast for either the Democrat or Republican parties.
}
