% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efficiency_gap.R
\name{vote_prep}
\alias{vote_prep}
\title{vote_prep takes in full_votes, of the object from year_baseline_data, and finds the total votes cast
for either of the major parties in each district. This number is joined to the primary data frame
so we can find the total votes, estimated or actually cast, for the state assembly cast by the whole state
which is necessary for the efficiency gap.}
\usage{
vote_prep(full_votes)
}
\description{
vote_prep takes in full_votes, of the object from year_baseline_data, and finds the total votes cast
for either of the major parties in each district. This number is joined to the primary data frame
so we can find the total votes, estimated or actually cast, for the state assembly cast by the whole state
which is necessary for the efficiency gap.
}
