% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efficiency_gap.R
\docType{data}
\name{wi_sa}
\alias{wi_sa}
\title{Establishes a data frame that has the State Assembly results for 2006-2020. This gets called
within the efficiency gap function.}
\format{
An object of class \code{data.frame} with 8 rows and 4 columns.
}
\usage{
wi_sa
}
\description{
Establishes a data frame that has the State Assembly results for 2006-2020. This gets called
within the efficiency gap function.
}
\keyword{datasets}
