% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline_data_generate.R
\name{year_baseline_data}
\alias{year_baseline_data}
\title{Retrieves election data for Wisconsin from the years 2000-2020}
\usage{
year_baseline_data(year, data)
}
\arguments{
\item{year}{A character vector identifying the requested year}

\item{data}{the primary data frame with everything for the state
as returned by generate data}
}
\value{
A list of eleven data frames, each of election data from Wisconsin and
year with the following columns
\itemize{
\item county
\item ward
\item office
\item district
\item total.votes
\item party
\item candidate
\item votes
\item contest_dem
\item contest_rep
\item year
}

Note that many of the fields may be an empty string
}
\description{
If given Wisconsin and a year, this function returns all
election data from the OpenElections github for that state and year
}
