% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data_final.R
\name{access_state_year}
\alias{access_state_year}
\title{Retrieves election data for Michigan and Wisconsin from the years 2000-2020}
\usage{
access_state_year(state, year)
}
\arguments{
\item{state}{A character vector representing either WI or MI}

\item{year}{A character vector identifying the requested year}
}
\value{
A dataframe of election data from the given state and year with the following columns
\itemize{
\item county
\item ward
\item office
\item district
\item total.votes
\item party
\item candidate
\item votes
\item contest_dem
\item contest_rep
\item year
}

Note that many of the fields may be an empty string
}
\description{
If given either Michigan or Wisconsin and a year, this function returns all
election data from the openelections github for that state and year
}
