% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleaning_uncontested.R
\name{total_vote_func}
\alias{total_vote_func}
\title{total_vote_func takes the filtered statewide information and finds}
\usage{
total_vote_func(x)
}
\arguments{
\item{x}{data frame created by open_elections_factory and generate_data}
}
\description{
total_vote_func takes the filtered statewide information and finds
}
